[#3] add directory-based config source
Implementation
--------------

Directory is scanned when the config source is created and
any files becomes a property (the file name being the key and the file
content being the value).
Scanning is *not* recursive.

Extension
---------

Add dir attribute to the config-source model to scan the directory and
create a ConfigSource from its content

Fraction
--------

Add missing module dependency to the config-api (that was causing a
ClassNotFoundError in uberJar)

Example
-------

Add openshift example that shows how to use the ConfigSource API with
OpenShift/Kubernetes Config Maps